+++ release_notes.md
@@ -14,7 +14,19 @@ This file tracks release notes for the loop client.
 
 ## Next release
 
-#### NewFeatures
+#### New Features
+
+##### Autoloop Swap Size
+* Autoloop can now be configured with custom swap size limits. Previously,
+autoloop would use the minimum/maximum swap amount set by the server (exposed 
+by the `loop terms` command) to decide on swap size. 
+* Setting a custom minimum swap amount is particularly useful for clients that 
+would like to perform fewer, larger swaps to save on fees. The trade-off when 
+setting a large minimum amount is that autoloop will wait until your channel is 
+at least the minimum amount below its incoming threshold amount before executing
+a swap, which may result in channels staying under the threshold for longer. 
+* These values can be set using the following command: `loop setparams --minamt={minimum in sats} --maxamt={maximum in sats}`.
+* The values set must fall within the limits set by the loop server. 
 
 #### Breaking Changes